feat: implement univ2nav pricing feed and update adapter for improved LP token pricing
