+++ projects/ichi/src/feeds/univ2nav.ts
@@ -1,103 +1,88 @@
-// scratch notes
+import { HandlerFactory } from './interface';
+import Big from 'big.js';
+import * as univ2Abi from '../abi/univ2';
+import { log } from '../utils/logger';
+import { CoreFeedSelector } from '../types/pricing';
 
-// NAV might be the same
-// however, underlying price might be different
-// coingecko vs defillama vs codex vs univ3
-// not to mention that they need to provide configuration per asset or could try to optimistically price everything
+// Handler name constant for Uniswap V2 LP NAV pricing
+export const UNIV2_NAV_HANDLER = 'univ2nav';
 
-/*
-mapping our price wiring:
-  1. find the first block of the window duration (which is the start of the window)
-  2. invoke the priceAsset function on that window
-  3. store the result of that computation into the price cache (aka: store the price for that window)
-  final output: redis timeseries that has one price per asset for that particular window
+// Official feed handler factory for Uniswap V2 LP tokens
+export const univ2NavFactory: HandlerFactory<typeof UNIV2_NAV_HANDLER> =
+  (resolve) =>
+  async ({ assetConfig, ctx }) => {
+    // Extract the univ2nav feed config from the asset config
+    const feedConfig = assetConfig.priceFeed as Extract<CoreFeedSelector, { kind: 'univ2nav' }>;
 
-  enrichment step:
-  1. for each window, we get the twa price for that particular duration by iterating over the price key in redis
-  2.
+    try {
+      // Create contract instance for the LP token (pool address is the asset key)
+      const poolAddress = ctx.asset.toLowerCase();
+      const lpContract = new univ2Abi.Contract(ctx.sqdRpcCtx, poolAddress);
 
-*/
+      // Cache key for pool configuration (use the asset key)
+      const poolConfigKey = poolAddress;
 
-/*
-order of operations for pricing:
-1. define the asset price feed config
-2.
+      // Try to get cached pool configuration first
+      let poolConfig = await ctx.handlerMetadataCache.get(UNIV2_NAV_HANDLER, poolConfigKey);
 
+      if (!poolConfig) {
+        // Get token addresses and decimals from the pool and cache them
+        const [token0Address, token1Address, lpDecimals] = await Promise.all([
+          lpContract.token0(),
+          lpContract.token1(),
+          lpContract.decimals(),
+        ]);
 
-1. for each flush interval (hourly, daily), get all the asset keys
-2. for each asset key, we need to price it somehow
+        poolConfig = {
+          token0Address: token0Address.toLowerCase(),
+          token1Address: token1Address.toLowerCase(),
+          lpDecimals: Number(lpDecimals.toString()),
+        };
 
+        // Cache the pool configuration
+        await ctx.handlerMetadataCache.set(UNIV2_NAV_HANDLER, poolConfigKey, poolConfig);
+      }
 
-// univ2 nav
-LP pricing: get reserves, get decimals, price both underlying assets, do some quick math
-  - decimals - cached. underlying asset price - cached. But these are all helpers, not inherent to the implementation.
-Swap pricing:
-- token0 amount scaled by decimals * price of token0 + token1 amount scaled by decimals * price of token1
+      // Get pool reserves and total supply using the contract methods
+      const [reservesData, totalSupply] = await Promise.all([
+        lpContract.getReserves(),
+        lpContract.totalSupply(),
+      ]);
 
-// univ3:
-asset: liquidity (L)
-user: owner of nft
-pricing logic:
-  for all in range positions:
-    - get state of pool (like upper/lower tick)
-    - get liquidity of pool
-    - get price of underlying assets
-    - do some quick math
-    - sum up the results
-    - return the result
+      const { _reserve0: reserve0, _reserve1: reserve1 } = reservesData;
 
-// simple asset:
-assset: address
-user: address
+      // Get prices for both underlying tokens using recursion
+      const [price0Result, price1Result] = await Promise.all([
+        resolve(
+          { assetType: feedConfig.token0.assetType, priceFeed: feedConfig.token0.priceFeed },
+          poolConfig.token0Address, // Use cached token address as asset key
+          ctx,
+        ),
+        resolve(
+          { assetType: feedConfig.token1.assetType, priceFeed: feedConfig.token1.priceFeed },
+          poolConfig.token1Address, // Use cached token address as asset key
+          ctx,
+        ),
+      ]);
 
-*/
+      // Calculate LP token price
+      // LP Price = (reserve0 * price0 + reserve1 * price1) / totalSupply
+      const token0Value = new Big(reserve0.toString())
+        .div(Math.pow(10, price0Result.metadata.decimals))
+        .mul(price0Result.price);
+      const token1Value = new Big(reserve1.toString())
+        .div(Math.pow(10, price1Result.metadata.decimals))
+        .mul(price1Result.price);
+      const totalPoolValue = token0Value.plus(token1Value);
 
-// OLD IMPLEMENTATION
-// // UNIV2 NAV PRICING
-// class UniV2Nav {
-//     constructor(
-//         private feed: PriceFeed,
-//         private ctx: any,
-//         private assetMap: Record<string, AssetKey>,
-//     ) { }
+      // Use cached LP token decimals
+      const lpTokenPrice = totalPoolValue.div(
+        new Big(totalSupply.toString()).div(Math.pow(10, poolConfig.lpDecimals)),
+      );
 
-//     /** Total pool USD value at time `atMs` */
-//     async poolValueUSD(lp: string, atMs: number): Promise<number> {
-//         const { token0, token1 } = await this.r.getPairTokens(lp);
-//         const [{ r0, r1 }, dec0, dec1] = await Promise.all([
-//             this.r.getReserves(lp, atMs),
-//             this.r.getDecimals(token0),
-//             this.r.getDecimals(token1),
-//         ]);
-
-//         const a0 = Number(r0) / 10 ** dec0;
-//         const a1 = Number(r1) / 10 ** dec1;
-
-//         // Map on-chain address -> your AssetKey, then -> CoinGecko id via feed config
-//         const key0 = this.assetMap[token0.toLowerCase()];
-//         const key1 = this.assetMap[token1.toLowerCase()];
-//         if (!key0 || !key1) throw new Error('Missing asset mapping for token0 or token1');
-
-//         const [p0, p1] = await Promise.all([
-//             this.feed.priceUSD(key0, atMs),
-//             this.feed.priceUSD(key1, atMs),
-//         ]);
-
-//         return a0 * p0 + a1 * p1;
-//     }
-
-//     /** Price of 1 LP token in USD at time `atMs` */
-//     async lpTokenPriceUSD(lp: string, atMs: number): Promise<number> {
-//         const [poolUsd, totalSupplyRaw] = await Promise.all([
-//             this.poolValueUSD(lp, atMs),
-//             this.r.getTotalSupply(lp, atMs),
-//         ]);
-//         // LP token has 18 decimals on typical UniV2-like pairs, but don’t assume
-//         const lpDecimals = 18; // read if you store it
-//         const totalSupply = Number(totalSupplyRaw) / 10 ** lpDecimals;
-//         if (totalSupply === 0) return 0;
-//         return poolUsd / totalSupply;
-//     }
-// }
-
-// // END UNIV2 NAV PRICING
+      return Number(lpTokenPrice.toString());
+    } catch (error) {
+      log.warn(`Failed to price Uniswap V2 LP token ${ctx.asset}:`, error);
+      return 0;
+    }
+  };
